requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
requests==2.31.0
jinja2==3.1.2
starlette==0.27.0
python-multipart==0.0.6
langchain-openai==0.1.0
mcp-use==0.1.0
